// frontend/src/ipc/generated/SearchResultView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SearchEntityKind } from "./SearchEntityKind";
import type { SearchNavigationTarget } from "./SearchNavigationTarget";
import type { SearchTextFragment } from "./SearchTextFragment";

export type SearchResultView = { entity_id: string, entity_kind: SearchEntityKind, title: string, title_fragments: Array<SearchTextFragment>, context_text: string, snippet_fragments: Array<SearchTextFragment>, navigation_target: SearchNavigationTarget, rank: number, };